remove view dto from domain
