--- domain/loginpage/login.go
@@ -1,7 +1,0 @@
-package loginpage
-
-type Login struct {
-	Name   string
-	Origin string
-	Error  string
-}
