+++ infrastructure/httpHandler/dashboard.go
@@ -30,6 +30,7 @@ func (dp DashboardPage) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	path := strings.TrimPrefix(r.URL.Path, "/dashboard")
 	requestId := getQueryValueOrNil(r.URL.Query(), "requestId")
 
-	dashBoardData := dp.DashboardService.GetDashboardData(path, requestId)
+	dashBoardData := dp.DashboardService.GenerateDashboardViewData(path, requestId)
+
 	htmlTemplate.PresentHtmlTemplate(w, "resources/dashboard.html", dashBoardData)
 }
